refactor(database.js): configuraçoes da database movidas para /configs/database.js
As configuraçoes de database foram movidas para outra pasta e arquivo, para melhorar a organizacao
do modulo
